+++ AnkiDroid/src/test/java/com/ichi2/libanki/sched/SchedTest.java
@@ -1183,7 +1183,7 @@ public class SchedTest extends RobolectricTest {
         Card c = getCard();
         // set a a fail delay of 4 seconds
         DeckConfig conf = col.getSched()._cardConf(c);
-        conf.getJSONObject("lapse").getJSONArray("delays").put(0, 1 / 15.0);
+        conf.getJSONObject("lapse").getJSONArray("delays").put((Integer) 0, 1 / 15.0);
         col.getDecks().save(conf);
         col.getSched().answerCard(c, 1);
         // the next card should be another review
